+++ GeoTools/process-spatialstatistics/src/main/java/org/geotools/process/spatialstatistics/storage/ShapefileFeatureInserter.java
@@ -455,7 +455,7 @@ public class ShapefileFeatureInserter implements IFeatureInserter {
         int retLength = 0;
         int tempSize = 0;
         for (int index = 1; index <= value.length(); index++) {
-            int asc = (int) value.charAt(index - 1);
+            int asc = value.charAt(index - 1);
             if (asc > 127) {
                 if (byte_len >= tempSize + 2) {
                     tempSize += 2;
